レビュー修正、新規課題 9.arrayList使う 10.double型にキャストをint変換関数に入れ込む 11.StringBufferの結合時にcalucurateをString型へ変換しないでそのまま入れる
3.resultを数値に、Rohrerの判定をjudgeに
13.HashMap

+++ src/kadai02/common/ChangeNumUtil.java
@@ -24,7 +24,7 @@ public class ChangeNumUtil {
 	public int changeStringToInt(String args) {
 
 		// 入力値をint型に変換し、変数に代入
-		int intChange = Integer.parseInt(args);
+		int intChange = (int)Double.parseDouble(args);
 
 		// int型に変換した値を返す
 		return intChange;
